pyproject.toml
[project]
name = "discord-potluck-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2026.1.post1",
]
